requirements.txt
pandas==2.1.0
numpy==1.25.0
scikit-learn==1.3.0
imbalanced-learn==0.11.0
matplotlib==3.7.2
seaborn==0.12.2
pyserial==3.5
joblib==1.3.1
python-dotenv==1.0.0
Pillow==10.0.0
jupyter==1.0.0
